Started new sample. Fixed bug.
