--- samples/Samples.Api/Design/Entities/Brand/Brand.loader.nox.yaml
@@ -1,47 +1,0 @@
-﻿#
-# Brand data loader definition
-#
-
-Name: BrandLoader
-
-Description: Loads brand data 
-
-Schedule:
-  Start: Daily at 2am UTC
-  Retry:
-    Limit: 10
-    DelaySeconds: 60
-    DoubleDelayLimit: 5
-
-LoadStrategy: 
-  Type: MergeNew
-  Columns: [CreateDate,EditDate]
-
-Target:
-  Entity: Brand
-  
-Messaging:
-  - MessagingProvider: AppServiceBus  
-
-Sources:
-  - DataSource: BusinessSystem
-    MinimumExpectedRecords: 30 
-    Query: >-
-      SELECT 
-           [BrandID] AS [Id]
-          ,[BrandName] AS [Name]
-          ,[BrandShortName] AS [ShortName]
-          , CASE
-              WHEN [CorporateVentureId] = 50 THEN 2
-              ELSE 1
-            END
-            AS [PlatformTenantId]
-          ,[BrandLogo] AS [Logo_1024x1024_Url]
-          ,[BrandLogo] AS [Logo_1024x768_Url]
-          ,[WebsiteUrl]
-          ,[MyRegusUrlKeyName] AS [CustomerPortalUrl]
-          ,[AndroidLink] AS [AndroidAppUrl]
-          ,[IOSLink] AS [AppleAppUrl]
-          ,[CreateDate]
-          ,[EditDate]
-      FROM dbo.Brand WITH (NOLOCK)
