--- samples/Samples.Api/Design/Entities/Country/Country.loader.nox.yaml
@@ -1,40 +1,0 @@
-﻿#
-# Country data loader definition
-#
-
-Name: CountryLoader
-
-Description: Loads country data 
-
-Schedule:
-  Start: Daily at 2am UTC
-  Retry:
-    Limit: 10
-    DelaySeconds: 60
-    DoubleDelayLimit: 5
-
-LoadStrategy: 
-  Type: MergeNew
-  Columns: [CreateDate,EditDate]
-
-Target:
-  Entity: Country
-  
-Messaging:
-  - MessagingProvider: InProcess    
-
-Sources:
-  - DataSource: BusinessSystem
-    MinimumExpectedRecords: 242 
-    Query: >-
-      SELECT 
-           [CountryISONumber] AS [Id]
-          ,[CountryName] AS [Name]
-          ,[CountryCodeISO2Alpha] AS [ISO_Alpha2]
-          ,[CountryCodeISO3Alpha] AS [ISO_Alpha3]
-          ,[CountryISONumber] AS [ISO_Numeric]
-          ,[CurrencyID] AS [CurrencyId]
-          ,[CreateDate]
-          ,[EditDate]
-      FROM dbo.Country WITH (NOLOCK)
-
