--- samples/Samples.Api/Design/Entities/PlatformTenant/PlatformTenant.entity.nox.yaml
@@ -1,24 +1,0 @@
-﻿#
-# Platform tenant entity definition
-#
-
-Name: PlatformTenant
-Description: Tenants of the platform
-Table: PlatformTenant
-
-Attributes:
-
-  - Name: Id
-    Description: The platform tenant's unique identifier 
-    IsPrimaryKey: true
-    Type: int
-    CanFilter: true
-    CanSort: true
-
-  - Name: Name
-    Description: The name of the platform tenant
-    IsRequired: true
-    Type: string
-    MaxWidth: 128
-    CanFilter: true
-    CanSort: true
