--- samples/Samples.Api/Design/Entities/PlatformTenant/PlatformTenant.loader.nox.yaml
@@ -1,24 +1,0 @@
-﻿#
-# Platform tenant data loader definition
-#
-
-Name: PlatformTenantLoader
-
-Description: Loads platform tenant data 
-
-Schedule:
-  Start: Daily at 2am UTC
-
-LoadStrategy: 
-  Type: MergeNew
-
-Target:
-  Entity: PlatformTenant
-
-Sources:
-  - DataSource: BusinessSystem
-    MinimumExpectedRecords: 2
-    Query: >-
-      SELECT 1 AS [Id], 'International Workplace Group (IWG) PLC' AS [Name]
-      UNION
-      SELECT 2 AS [Id], 'Worka Inc.' AS [Name]
