--- samples/Samples.Api/Design/Entities/Workplace/Workplace.api.nox.yaml
@@ -1,43 +1,0 @@
-﻿#
-# Workplace api definitions
-#
-
-Name: WorkplaceApi
-Description: Workplace entity controller and endpoints
-
-Routes:
-
-  - Name: Workplaces
-    HttpVerb: GET
-    Description: Returns all workplaces
-    TargetUrl: odata/Workplaces
-    Responses:
-      - Type: Workplace
-        IsCollection: true 
-    
-  - Name: Workplace/{Id}
-    HttpVerb: GET
-    Description: Returns a specific workplace
-    Parameters:
-      - Name: Id
-        Type: int
-    TargetUrl: odata/Workplaces({Id})
-    Responses:
-      - Type: Workplace
-        IsCollection: false
-
-  - Name: WorkplacesWithPagination
-    HttpVerb: GET
-    Description: Returns workplaces with pagination
-    Parameters:
-      - Name: PageNumber
-        Type: int
-        Default: 1
-      - Name: PageSize
-        Type: int
-        Default: 20
-    TargetUrl: odata/Workplaces?$skip={PageNumber}&$top={PageSize}
-    Responses:
-      - Type: Workplace
-        IsCollection: true 
-
