--- samples/Samples.Api/Design/Entities/Workplace/Workplace.entity.nox.yaml
@@ -1,67 +1,0 @@
-﻿#
-# Workplace entity definition
-#
-
-Name: Workplace
-Description: Information on places where people can work
-PluralName: Workplaces
-Table: Workplace
-Schema: dbo
-
-RelatedParents: [Brand,Country]
-
-Attributes:
-
-  - Name: Id
-    Description: The Workplace's unique identifier 
-    IsPrimaryKey: true
-    Type: int
-    CanFilter: true
-    CanSort: true
-
-  - Name: Name
-    Description: The name of the Workplace
-    IsRequired: true
-    Type: string
-    MaxWidth: 128
-    CanFilter: true
-    CanSort: true
-
-  - Name: Latitude
-    Description: The the north–south position of the workplace's point on the surface of the Earth
-    Type: decimal
-    MaxWidth: 18
-    Precision: 15
-
-  - Name: Longitude
-    Description: The the east-west position of the workplace's point on the surface of the Earth
-    Type: decimal
-    MaxWidth: 18
-    Precision: 15
-
-  #- Name: OpeningTime
-    #Description: The opening time for the workplace
-    #Type: time
-    #IsRequired: false
-    #Default: '08:30'
-    #DefaultFromParents: [AdministrativeArea,Country]
-
-  #- Name: ClosingTime
-    #Description: The closing time for the workplace
-    #Type: time
-    #IsRequired: false
-    #Default: '17:30'
-    #DefaultFromParents: [AdministrativeArea,Country]
-
-  #- Name: Is24Hour
-    #Type: bool
-    #Formula: AND(OpeningTime="00:00:00",ClosingTime="23:59:59")
-
-  #- Name: IsPublicHoliday
-    #Description: Is it a public holiday
-    #Type: bool
-    #Default: false
-    #DefaultFromParents: [AdministrativeArea,Country]
-    #IsTemporalOnly: true
-
-
