--- samples/Samples.Api/Design/Entities/Workplace/Workplace.loader.nox.yaml
@@ -1,38 +1,0 @@
-﻿#
-# Workplace data loader definition
-#
-
-Name: WorkplaceLoader
-
-Description: Loads workplace data 
-
-Schedule:
-  Start: Daily at 2am UTC
-  Retry:
-    Limit: 10
-    DelaySeconds: 60
-    DoubleDelayLimit: 5
-
-LoadStrategy: 
-  Type: MergeNew
-  Columns: [CreateDate,EditDate]
-
-
-Target:
-  Entity: Workplace
-
-Sources:
-  - DataSource: BusinessSystem
-    MinimumExpectedRecords: 3000
-    Query: >-
-      SELECT  [CentreNumber] AS [Id]
-              ,[CentreName] AS [Name]
-              ,[Latitude]
-              ,[Longitude]
-              ,[BrandID] AS [BrandId]
-              ,[Country].[CountryISONumber] AS [CountryId]
-              ,[Centre].[CreateDate] AS [CreateDate]
-              ,[Centre].[EditDate] AS [EditDate]
-      FROM 		  dbo.[Centre] WITH (NOLOCK)
-      LEFT JOIN	dbo.[Country] ON [Country].[CountryID]=[Centre].[CountryID]
-      WHERE 		[Centre].[IsCurrentVersion]=1
