+++ samples/Samples.Api/Design/SampleCurrency.service.nox.yaml
@@ -2,7 +2,7 @@
 # Service definition
 #
 
-Name: Workplace
+Name: SampleCurrency
 
 Description: Workplace Microservice
